+++ daikon-spring/hystrix-utils/src/main/java/org/talend/daikon/hystrix/CommandHelper.java
@@ -9,6 +9,7 @@ import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.talend.daikon.exception.TalendRuntimeException;
+import org.talend.daikon.hystrix.processors.InternalErrorCodes;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.netflix.hystrix.HystrixCommand;
