--- daikon-spring/hystrix-utils/src/main/java/org/talend/daikon/hystrix/HttpSecurityProvider.java
@@ -1,9 +1,0 @@
-package org.talend.daikon.hystrix;
-
-import org.apache.http.client.methods.HttpRequestBase;
-
-public interface HttpSecurityProvider {
-
-    void addSecurityHeaders(HttpRequestBase request);
-
-}
